lib/rdma: align device capability limits with firmware array bounds
Replace inflated DEFAULT_MAX_VALUE (1M) defaults with 1024 for max
QPs, CQs, AH, SRQ, ucontexts, and QP WRs. Clamp PTS-reported values
in rdma_query_device_cap() to firmware limits. Widen CQ index from
uint8_t to uint16_t to fix always-false bounds checks. Fix
pd_entry.mr_pool sizing from RDMA_MAX_PD to RDMA_MAX_MR.

+++ lib/rdma/rdma_cq.h
@@ -5,9 +5,9 @@
 #ifndef __RDMA_CQ_H__
 #define __RDMA_CQ_H__
 
-uint64_t rdma_cq_get(uint8_t index);
+uint64_t rdma_cq_get(uint16_t index);
 void rdma_cq_put(void);
-int rdma_cq_insert(uint64_t addr, uint8_t index);
-int rdma_cq_remove(uint8_t index);
+int rdma_cq_insert(uint64_t addr, uint16_t index);
+int rdma_cq_remove(uint16_t index);
 
 #endif /* __RDMA_CQ_H__ */
